Cambios de formato y demás actualizaciones.
Fuera de las practicas

Le cambie el nombre a cada carpeta con el fin de que se vea mas ordenado todo (practica 1, 2,...., 10), asi como agregar la carpeta de csv para almacenar los distintos csv y todo este mas ordenado.

Practica 1

Actualice la localización del DataSet, dado que implemente una nueva carpeta donde se almacenaran los csv.

Practica 2

La lectura ha sido actualizada para que se lea el DataSet en la carpeta csv, así como a la hora de hacer el to_csv a la misma carpeta en el mismo doc (dado que se reemplaza)

Practica 3

La lectura ha sido actualizada así como donde se guarda el nuevo archivo, siendo en csv/Analyzed DataSet Departments.csv

Practica 4

Decidí en vez de hacer los plots de categorías, mas bien de departamentos, por lo que se eliminaron las primeras dos funciones que se encargaban de categorizar (que en realidad no ocupaba dado que ya existen las categorías), la razón por la que no quise dejarlo así con las categorías, es que las categorías del DataSet original, es decir, la columna Job Category, tiene demasiadas categorías, entonces a la hora de los plots es demasiado que ni se aprecia de buena forma, es exceso de información en mi opinión, por lo que decidi manejar los plots con los departamentos unicamente, que de hecho en el PiePlot tampoco se aprecia del todo bien, pero aun así se aprecia mejor que con categorías, hablo de categorías del dataset original, las categorías que cree yo manualmente previamente funcionaban bien para el pieplot, pero en realidad se supone que debería de trabajar con el dataset, por lo tanto utilizar las categorías del dataset. En resumen ahora es sobre departamentos.

Practica 5

La lectura del csv

Practica 6

La lectura del csv

Practica 7

La lectura del csv

Practica 8

La lectura del csv

Practica 9

La lectura del csv

Practica 10

La lectura del csv
