--- Categorized Departments.csv
@@ -1,62 +1,0 @@
-Departamento de Trabajo,Salario Total,Cantidad de Sueldos,Promedio Salario,Salario Maximo,Categoria
-ADMIN FOR CHILDREN'S SVCS,9894208.9907,122,81100.0736942623,132902.0,
-ADMIN TRIALS AND HEARINGS,2025179.715,31,65328.37790322581,144000.0,Administracion
-BOROUGH PRESIDENT-BRONX,399025.5,7,57003.642857142855,75442.0,Administracion
-BRONX COMMUNITY BOARD #7,55798.0,1,55798.0,55798.0,Administracion
-BRONX COMMUNITY BOARD #8,75442.0,1,75442.0,75442.0,Administracion
-BRONX DISTRICT ATTORNEY,5731899.33,64,89560.92703125,195900.0,Justicia
-BUSINESS INTEGRITY COMMISSION,338232.0,5,67646.4,112500.0,Regulacion y Cumplimiento
-CAMPAIGN FINANCE BOARD,2292370.0,28,81870.35714285714,226082.5,Regulacion y Cumplimiento
-CIVILIAN COMPLAINT REVIEW BD,555262.7,7,79323.24285714285,144878.5,Justicia
-CONFLICTS OF INTEREST BOARD,69614.0,1,69614.0,69614.0,Regulacion y Cumplimiento
-CONSUMER AND WORKER PROTECTION,1184032.6,21,56382.50476190477,89810.0,Regulacion y Cumplimiento
-CULTURAL AFFAIRS,231553.2923,6,38592.215383333336,61486.0,Cultura y Arte
-DEPARTMENT FOR THE AGING,1135249.6306,23,49358.679591304346,112000.0,Servicios Sociales
-DEPARTMENT OF BUILDINGS,543326.5,7,77618.07142857143,109961.0,Construccion
-DEPARTMENT OF BUSINESS SERV.,1376683.5,19,72457.02631578948,97275.0,Servicios Empresariales
-DEPARTMENT OF CITY PLANNING,2447820.0,28,87422.14285714286,125000.0,Planificacion Urbana
-DEPARTMENT OF CORRECTION,8851737.91785,99,89411.49411969697,212500.0,Seguridad Publica
-DEPARTMENT OF FINANCE,69792.99,2,34896.495,69735.0,Finanzas
-DEPARTMENT OF INVESTIGATION,1700592.0,22,77299.63636363637,135000.0,Justicia
-DEPARTMENT OF PROBATION,523565.5,6,87260.91666666667,132095.0,Seguridad Publica
-DEPARTMENT OF SANITATION,641765.465,10,64176.5465,155432.5,Servicios Publicos
-DEPARTMENT OF TRANSPORTATION,19594296.12165,257,76242.39736050583,184335.5,Transporte
-DEPT OF CITYWIDE ADMIN SVCS,2737035.0,39,70180.38461538461,197500.0,Administracion
-DEPT OF DESIGN & CONSTRUCTION,14014089.5,181,77425.908839779,117818.0,Construccion
-DEPT OF ENVIRONMENT PROTECTION,41482408.133,474,87515.62897257385,235000.0,Medio Ambiente
-DEPT OF HEALTH/MENTAL HYGIENE,36425443.31555,586,62159.45958284983,174081.5,Salud
-DEPT OF PARKS & RECREATION,2726112.0992,44,61957.09316363637,128000.0,Parques y Recreacion
-DEPT OF RECORDS & INFO SERVICE,321239.04945,6,53539.841575,94715.0,
-DEPT OF YOUTH & COMM DEV SRVS,4904378.9898500005,77,63693.23363441559,164902.0,Servicios Comunitarios
-DEPT. OF HOMELESS SERVICES,2534961.22,37,68512.46540540541,181495.5,Servicios Sociales
-DISTRICT ATTORNEY KINGS COUNTY,168153.0,3,56051.0,67983.0,Justicia
-DISTRICT ATTORNEY RICHMOND COU,1106769.54985,17,65104.09116764706,132500.0,Justicia
-EQUAL EMPLOY PRACTICES COMM,63549.5,1,63549.5,63549.5,Recursos Humanos
-FINANCIAL INFO SVCS AGENCY,4000203.6,49,81636.8081632653,145000.0,Finanzas
-FIRE DEPARTMENT,6287276.6731,75,83830.35564133334,170000.0,Seguridad Publica
-HOUSING PRESERVATION & DVLPMNT,4947190.0,61,81101.47540983607,200000.0,Vivienda
-HRA/DEPT OF SOCIAL SERVICES,24528187.245,318,77132.66429245283,152500.0,Servicios Sociales
-HUMAN RIGHTS COMMISSION,1466535.0,19,77186.05263157895,115465.5,Recursos Humanos
-LANDMARKS PRESERVATION COMM,66272.0,1,66272.0,66272.0,Cultura y Arte
-LAW DEPARTMENT,4343576.0073,77,56410.07801688311,152721.0,Justicia
-MANHATTAN COMMUNITY BOARD #1,95000.0,1,95000.0,95000.0,Administracion
-MANHATTAN COMMUNITY BOARD #10,87500.0,1,87500.0,87500.0,Administracion
-MAYORS OFFICE OF CONTRACT SVCS,783043.5,12,65253.625,187500.0,Administracion
-MUNICIPAL WATER FIN AUTHORITY,179413.0,1,179413.0,179413.0,Finanzas
-NYC DEPT OF VETERANS' SERVICES,291474.46335,5,58294.89267,97500.0,Servicios Sociales
-NYC EMPLOYEES RETIREMENT SYS,1293364.5,14,92383.17857142857,140000.0,Recursos Humanos
-NYC HOUSING AUTHORITY,11814265.31,179,66001.48217877095,156348.5,Vivienda
-NYC POLICE PENSION FUND,341117.5,4,85279.375,119204.0,Recursos Humanos
-OFF OF PAYROLL ADMINISTRATION,495989.0,7,70855.57142857143,120465.5,Administracion
-OFFICE OF CRIMINAL JUSTICE,539780.5,7,77111.5,89350.0,Justicia
-OFFICE OF EMERGENCY MANAGEMENT,862545.25,14,61610.375,102500.0,Seguridad Publica
-OFFICE OF LABOR RELATIONS,481816.5,5,96363.3,144878.5,Recursos Humanos
-OFFICE OF MANAGEMENT & BUDGET,6062426.0,73,83046.93150684932,222924.0,Administracion
-OFFICE OF THE COMPTROLLER,3661593.2,37,98961.97837837838,205000.0,Administracion
-POLICE DEPARTMENT,2098983.74,29,72378.74965517242,187500.0,Seguridad Publica
-PRESIDENT BOROUGH OF MANHATTAN,140000.0,1,140000.0,140000.0,Administracion
-PUBLIC ADMINISTRATOR-NEW YORK,111864.0,2,55932.0,59225.5,Administracion
-TAX COMMISSION,165659.0,2,82829.5,121601.5,Impuestos
-TAXI & LIMOUSINE COMMISSION,5010134.08645,73,68631.9737869863,140000.0,Transporte
-TEACHERS RETIREMENT SYSTEM,218385.5,2,109192.75,124035.5,Recursos Humanos
-TECHNOLOGY & INNOVATION,409208.0,6,68201.33333333333,107500.0,Tecnologia e Innovacion
