+++ csv/Analyzed DataSet Departments.csv
@@ -1,0 +1,65 @@
+Departamento de Trabajo,Salario Total,Cantidad de Sueldos,Promedio Salario,Salario Maximo
+ADMIN FOR CHILDREN'S SVCS,3238294.4907,41,78982.79245609757,132095.0
+ADMIN TRIALS AND HEARINGS,1553094.33,20,77654.71650000001,145000.0
+BOROUGH PRESIDENT-BRONX,287429.5,5,57485.9,75442.0
+BRONX COMMUNITY BOARD #7,55798.0,1,55798.0,55798.0
+BRONX COMMUNITY BOARD #8,75442.0,1,75442.0,75442.0
+BRONX DISTRICT ATTORNEY,6953167.83,80,86914.597875,195900.0
+BUSINESS INTEGRITY COMMISSION,213326.5,3,71108.83333333333,112500.0
+CAMPAIGN FINANCE BOARD,2279870.0,28,81423.92857142857,226082.5
+CIVILIAN COMPLAINT REVIEW BD,334942.2,5,66988.44,110117.0
+COMMISSION ON RACIAL EQUITY,499250.0,6,83208.33333333333,114350.0
+CONFLICTS OF INTEREST BOARD,69614.0,1,69614.0,69614.0
+CONSUMER AND WORKER PROTECTION,740681.0,11,67334.63636363637,87225.5
+CULTURAL AFFAIRS,231553.2923,6,38592.215383333336,61486.0
+DEPARTMENT FOR THE AGING,1125798.3257,17,66223.4309235294,112000.0
+DEPARTMENT OF BUILDINGS,210076.0,3,70025.33333333333,83899.5
+DEPARTMENT OF CITY PLANNING,1538309.0,16,96144.3125,123277.5
+DEPARTMENT OF CORRECTION,6089031.41785,69,88246.83214275361,212500.0
+DEPARTMENT OF FINANCE,489017.99,6,81502.99833333334,147500.0
+DEPARTMENT OF INVESTIGATION,1158054.5,14,82718.17857142857,135000.0
+DEPARTMENT OF PROBATION,327901.5,3,109300.5,119204.0
+DEPARTMENT OF SANITATION,363219.46499999997,6,60536.57749999999,155432.5
+DEPARTMENT OF TRANSPORTATION,18256590.88165,236,77358.43593919493,200000.0
+DEPT OF CITYWIDE ADMIN SVCS,4826250.82,66,73125.01242424244,197500.0
+DEPT OF DESIGN & CONSTRUCTION,15645269.5,196,79822.80357142857,124718.5
+DEPT OF ENVIRONMENT PROTECTION,42940423.40185,497,86399.24225724346,235000.0
+DEPT OF HEALTH/MENTAL HYGIENE,36620351.56255,566,64700.267778356894,167406.5
+DEPT OF PARKS & RECREATION,1778348.4806,32,55573.39001875,95000.0
+DEPT OF RECORDS & INFO SERVICE,321210.7,5,64242.14,94715.0
+DEPT OF YOUTH & COMM DEV SRVS,3562348.33985,52,68506.69884326923,144878.5
+DEPT. OF HOMELESS SERVICES,3178688.42,45,70637.52044444444,181495.5
+DISTRICT ATTORNEY KINGS COUNTY,98170.0,2,49085.0,50000.0
+DISTRICT ATTORNEY RICHMOND COU,987276.5,15,65818.43333333333,132500.0
+EQUAL EMPLOY PRACTICES COMM,138991.5,2,69495.75,75442.0
+FINANCIAL INFO SVCS AGENCY,4220445.6,52,81162.41538461538,145000.0
+FIRE DEPARTMENT,5184254.63205,63,82289.75606428571,159969.5
+HOUSING PRESERVATION & DVLPMNT,3550893.0,47,75550.91489361702,198500.0
+HRA/DEPT OF SOCIAL SERVICES,26200065.045,344,76162.97978197676,165000.0
+HUMAN RIGHTS COMMISSION,960528.0,12,80044.0,115465.5
+LAW DEPARTMENT,3489141.3929,65,53679.09835230769,130000.0
+MANHATTAN COMMUNITY BOARD #10,87500.0,1,87500.0,87500.0
+MANHATTAN COMMUNITY BOARD #12,20.19,1,20.19,20.19
+MANHATTAN COMMUNITY BOARD #5,55000.0,1,55000.0,55000.0
+MANHATTAN COMMUNITY BOARD #8,50000.0,1,50000.0,50000.0
+MAYORS OFFICE OF CONTRACT SVCS,1227807.5,17,72223.9705882353,187500.0
+MUNICIPAL WATER FIN AUTHORITY,179413.0,1,179413.0,179413.0
+NYC DEPT OF VETERANS' SERVICES,25.46335,1,25.46335,25.46335
+NYC EMPLOYEES RETIREMENT SYS,988393.3342,13,76030.25647692308,140000.0
+NYC FIRE PENSION FUND,106243.0,2,53121.5,53121.5
+NYC HOUSING AUTHORITY,10113489.98,151,66976.75483443709,162500.0
+NYC POLICE PENSION FUND,167665.0,2,83832.5,114385.0
+OFF OF PAYROLL ADMINISTRATION,646142.0,9,71793.55555555556,120465.5
+OFFICE OF CRIMINAL JUSTICE,157793.5,2,78896.75,104350.0
+OFFICE OF EMERGENCY MANAGEMENT,1434704.75,22,65213.85227272727,102500.0
+OFFICE OF LABOR RELATIONS,494249.0,5,98849.8,144878.5
+OFFICE OF MANAGEMENT & BUDGET,2708549.5,33,82077.25757575757,185886.0
+OFFICE OF RACIAL EQUITY,912300.0,12,76025.0,114350.0
+OFFICE OF THE COMPTROLLER,3891609.25,40,97290.23125,205000.0
+POLICE DEPARTMENT,1602407.52,22,72836.70545454546,187500.0
+PRESIDENT BOROUGH OF MANHATTAN,140000.0,1,140000.0,140000.0
+PUBLIC ADMINISTRATOR-NEW YORK,111864.0,2,55932.0,59225.5
+TAX COMMISSION,165659.0,2,82829.5,121601.5
+TAXI & LIMOUSINE COMMISSION,2596230.5,38,68321.8552631579,117500.0
+TEACHERS RETIREMENT SYSTEM,226866.4,3,75622.13333333333,124000.0
+TECHNOLOGY & INNOVATION,802861.0,11,72987.36363636363,107500.0
